define data model for Hooks Service
