+++ hooks-checking-service/src/main/java/basilisk/hooksCheckingService/HooksCheckingServiceApplication.java
@@ -1,7 +1,6 @@
 package basilisk.hooksCheckingService;
 
-import basilisk.hooksCheckingService.domain.HooksRepos.GitBranchRepo;
-import basilisk.hooksCheckingService.domain.HooksRepos.GitRepo;
+import basilisk.hooksCheckingService.domain.HubRepos.GitBranchHubRepo;
 import basilisk.hooksCheckingService.messaging.HookMessageSender;
 import basilisk.hooksCheckingService.messaging.HookMessageSenderDumpImpl;
 import basilisk.hooksCheckingService.repositories.GitHookRepository;
@@ -11,15 +10,13 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-import java.io.IOException;
-
 @SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
 public class HooksCheckingServiceApplication {
 
     public static void main(String[] args) throws Exception {
         SpringApplication.run(HooksCheckingServiceApplication.class, args);
 
-        GitBranchRepo repo=new GitBranchRepo("Secure-Bank-And-ATms","Shaheen47",false,"","master");
+        GitBranchHubRepo repo=new GitBranchHubRepo("Secure-Bank-And-ATms","Shaheen47",false,"","master");
         GitHookRepository gitHookRepository=new GitHookRepositoryImpl();
         HookMessageSender hookMessageSender=new HookMessageSenderDumpImpl();
