--- hooks-checking-service/src/main/java/basilisk/hooksCheckingService/domain/HooksRepos/DockerRepo.java
@@ -1,4 +1,0 @@
-package basilisk.hooksCheckingService.domain.HooksRepos;
-
-public class DockerRepo {
-}
