+++ hooks-checking-service/src/main/java/basilisk/hooksCheckingService/domain/HubRepos/DockerRepo.java
@@ -1,0 +1,4 @@
+package basilisk.hooksCheckingService.domain.HubRepos;
+
+public class DockerRepo {
+}
